+++ src/extensions/filter-control/bootstrap-table-filter-control.js
@@ -130,7 +130,8 @@
                 {
                     field: $(this).closest('[data-field]').data('field'),
                     value: $(this).val(),
-                    position: getCursorPosition($(this).get(0))
+                    position: getCursorPosition($(this).get(0)),
+                    hasFocus: $(this).is(":focus")
                 });
         });
     };
@@ -142,6 +143,8 @@
             searchControls = getCurrentSearchControls(that);
 
         if (that.options.valuesFilterControl.length > 0) {
+            //  Callback to apply after settings fields values
+            var fieldToFocusCallback = null;
             header.find(searchControls).each(function (index, ele) {
                 field = $(this).closest('[data-field]').data('field');
                 result = $.grep(that.options.valuesFilterControl, function (valueObj) {
@@ -150,9 +153,24 @@
 
                 if (result.length > 0) {
                     $(this).val(result[0].value);
-                    setCursorPosition($(this).get(0), result[0].position);
+                    if(result[0].hasFocus) {
+                        // set callback if the field had the focus.
+                        fieldToFocusCallback = (function(fieldToFocus, carretPosition) {
+                            // Closure here to capture the field and cursor position 
+                            var closedCallback = function() {
+                                fieldToFocus.focus();
+                                setCursorPosition(fieldToFocus, carretPosition);
+                            };
+                            return closedCallback;
+                        })($(this).get(0), result[0].position);
+                    }
                 }
             });
+
+            // Callback call.
+            if(fieldToFocusCallback != null) {
+                fieldToFocusCallback();
+            }
         }
     };
 
@@ -224,6 +242,8 @@
                 }
             }
         });
+
+        that.trigger('created-controls');
     };
 
     var escapeID = function(id) {
@@ -438,6 +458,9 @@
         onColumnSearch: function (field, text) {
             return false;
         },
+        onCreatedControls: function() {
+            return true;
+        },
         filterShowClear: false,
         alignmentSelectControlOptions: undefined,
         filterTemplate: {
@@ -468,7 +491,8 @@
     });
 
     $.extend($.fn.bootstrapTable.Constructor.EVENTS, {
-        'column-search.bs.table': 'onColumnSearch'
+        'column-search.bs.table': 'onColumnSearch',
+        'created-controls.bs.table': 'onCreatedControls'
     });
 
     $.extend($.fn.bootstrapTable.defaults.icons, {
@@ -670,13 +694,32 @@
                 table = header.closest('table'),
                 controls = header.find(getCurrentSearchControls(that)),
                 search = that.$toolbar.find('.search input'),
+                hasValues = false,
                 timeoutId = 0;
 
             $.each(that.options.valuesFilterControl, function (i, item) {
+                hasValues = hasValues ? true : item.value !== '';
                 item.value = '';
             });
 
             setValues(that);
+
+             // clear cookies once the filters are clean
+             clearTimeout(timeoutId);
+             timeoutId = setTimeout(function () {
+                 if (cookies && cookies.length > 0) {
+                     $.each(cookies, function (i, item) {
+                         if (that.deleteCookie !== undefined) {
+                             that.deleteCookie(item);
+                         }
+                     });
+                 }
+             }, that.options.searchTimeOut);
+
+            //If there is not any value in the controls exit this method
+            if(!hasValues) {
+                return;
+            }
 
             // Clear each type of filter if it exists.
             // Requires the body to reload each time a type of filter is found because we never know
@@ -696,22 +739,10 @@
             if (that.options.sortName !== table.data('sortName') || that.options.sortOrder !== table.data('sortOrder')) {
                 var sorter = header.find(sprintf('[data-field="%s"]', $(controls[0]).closest('table').data('sortName')));
                 if (sorter.length > 0) {
-                    that.onSort(table.data('sortName'), table.data('sortName'));
+                    that.onSort({type: 'keypress', currentTarget: sorter});
                     $(sorter).find('.sortable').trigger('click');
                 }
             }
-
-            // clear cookies once the filters are clean
-            clearTimeout(timeoutId);
-            timeoutId = setTimeout(function () {
-                if (cookies && cookies.length > 0) {
-                    $.each(cookies, function (i, item) {
-                        if (that.deleteCookie !== undefined) {
-                            that.deleteCookie(item);
-                        }
-                    });
-                }
-            }, that.options.searchTimeOut);
         }
     };