+++ CHANGELOG.md
@@ -1,6 +1,34 @@
 ChangeLog
 ---------
 
+### 1.13.0
+
+- feat(js): rewrite bootstrap-table to ES6
+- feat(locale): add fi-FI.js locale
+- feat(build): use babel instead grunt
+- feat(filter-control): add `created-controls.bs.table` event to filter-control
+- feat(export extension): rewrite export extension to ES6
+- feat(export extension): export extension support bootstrap v4
+- feat(export extension): add `exportTable` method
+- feat(toolbar extension): rewrite toolbar extension to ES6
+- feat(toolbar extension): toolbar extension supports bootstrap v4
+- feat(toolbar extension): add server sidePagination support
+- feat(resizable extension): new resizable extension version 2.0.0
+- feat(editable extension): allow different x-editable configuration per table row
+- feat(addrbar extension): add addrbar extension
+- fix(js): fix #1769, improve check/uncheck methods
+- fix(js): fix #1983, cookie with pageNumber and searchText bug
+- fix(js): fix #2485, selections bugs
+- fix(js): fix #2545, customSearch support data attribute
+- fix(js): fix #3696, can't search data with formatter
+- fix(js): fix #4081, getRowByUniqueId error when row unique id is undefined
+- fix(js): fix older bootstrap version bug
+- fix(css): fix #1848, remove toolbar line-height
+- fix(css): limit fullscreen CSS rule scope
+- fix(editable extension): fix #1819, #2072, editable formatter bug
+- fix(extension): fix #3720, fix #3682, bug with export extension together
+- fix(extension): remove lick-edit-row and flat-json extensions
+
 ### 1.12.2
 
 - fix(js): fix #3656, toggle icon typo release error